Drop dpkg docs removal logic
We install minbase with debootstrap which should already be very
minimal, so if there's docs in there let's assume they're critical.

We already specify --no-install-recommends for apt which should
exclude docs packages as they're optional so there's no need for
dpkg specific overrides to exclude docs on top of that.