Reformat run commands to have "--option", "value" pairs on the same line
We were already doing this for the majority of invocations, but there
were some exceptions. Since this makes it much easier to mentally
split the command into logical parts, let's do this everywhere.

To make the command stand out a bit, add one space of seperation between
"[" and the first argument and between the last argument and "]".