Rework exception handling yet again
e25d746f9d7668edc5134cde4ec381c8b2da0136 removed printing of the exception
backtrace, except if --debug is used. (Later
afb709356fed78a1b6994e72d651e92f354be333 added back printing of the exception
message, and b89db00359ad28e90a1923f3b483b63e7f47829a changed it to print the
class name, but the general approach of not printing the backtrace remained.)
This is not very useful.

My motivating use case was an error introduced in one of my patches. A ValueError
was thrown by subprocess.run(). This would be trivial to fix except that
without the backtrace I had no idea what was going on.

We shouldn't print an exception if the failure is "expected". When we call
die(), then the code was written to expect a failure at this point and we print
an error message and we don't need to print the backtrace. So far good. But if
a different exception type is thrown, then most likely we made a programmatic
error and called something with wrong arguments, or the environment
unexpectedly prevents us from doing something that shouldn't fail. In that
case, just the exception (class or string or whatever) is not enough to
diagnose the issue, because exception names and messages are generic, and
without a backtrace won't know where exactly the failure happened. Telling the
user to rerun with --debug is not a good solution, because they might not be
able to recreate the failure.

The code is reworked to use MkosiError for our own exceptions and we only print
the error message (unless --debug is used). Other exceptions print the
backtrace, on the assumption that this is a failure that should be reported to
developers and fixed. (Or if a developer is using mkosi, they fix it themselves
;)). If it turns out that mkosi fails in a way that prints a backtrace in
normal use, then we can always add special handling for that case. This way, we
won't get backtraces in normal use, but we will get them if there's something
to fix in the code.

'raise MkosiError from e' is used so that we get chained backtraces.

I added 'type: ignore' because of the following mypy error:
  tests/test_backend.py:43:12: error: Exception type must be derived from BaseException  [misc]
        except type1 as e:
               ^~~~~
I don't think we can express that type1 must be a subclass of Exception.