Use an entry point instead of a script
The use of a custom script to launch mkosi has several shortcomings:

1. It translates `$PKEXEC_UID` into various `$SUDO_`* variables. That
   translation is lost if mkosi is run as a zipapp or as a module with
   `pkexec python3 -m mkosi`.
2. It has to muck around with `$PYTHONPATH` to launch mkosi directly
   from its source code.
3. It is tied to setuptools and has no equivalent in the standard
   project metadata (as in pyproject.toml).
4. And it also has to find the correct interpreter and maybe muck around
   with `$PYTHONPATH` to launch mkosi from its installation, which may
   be on the system, in the user's sitedir or in a virtualenv, as a
   standard or an editable installation.

Point 1 is not needed anymore now that `mkosi.backend.current_user()` is
used everywhere.

Point 2 can be done by directly running mkosi as a module from its
sources with:

- `[sudo|pkexec] python3 -m mkosi` inside the project directory
- `[sudo] PYTHONPATH=<PROJECT_PATH> python3 -m mkosi` from anywhere
- `[sudo|pkexec] <PROJECT_PATH>/bin/mkosi` from anywhere. The script is
  now a simple shim that is only used for this use case and not
  installed anymore.

Point 3 and 4 are better supported by using a `"console_scripts"` entry
point when installed. However, it is not possible anymore to run a
user-installed mkosi with sudo/pkexec.