Support additional keyrings on Debian derivatives
If `mkosi` is run with additional repositories (i.e.: `mkosi.reposdir`), the
specified repositories should include a path to their own keyring used for
verification of the repository. Normally the keyrings are stored under
`/etc/apt/keyrings` although `/usr/share/keyrings` could also be extended.

If the directory `/etc/apt/keyrings` is found then it will be bind-mounted
into place for `apt-get` to use. The directory would normally be populated
with `mkosi.skeleton` or similar.

If the directory `/usr/share/keyrings` is found then the contents of the host
directory `/usr/share/keyrings` and the skeleton directory are combined.
This prevents issues during bootstrap where required keys are not within
the skeleton directory.