Add support for PackageDirectories
This change supports installing package files from within a directory. Only
direct files are supported (no recursive lookups). The intent is to install
downloaded packages that are not available from a repository, and to directly
install built packages after running the build script.

Gentoo and OpenSuSE are not supported due to the way their package managers
operate.