gentoo: default to btrfs
Given that mkosi is *bespoken* and since it leverages features from
btrfs in some configurations we switch to btrfs by default.

Gentoo doesn't actually care one way or the other what the undelying fs
is and ext4, previous default, was in fact an arbitrarily choice (a
copy-pasted from some other distro's module).

Signed-off-by: Paymon MARANDI <darwinskernel@gmail.com>