config: rewrite config_make_list_parser
This removes the unescape option in favour of always using shlex. We also split
on everything shlex splits, not just whitespace, so that we get a fuller token
stream, allowing for arbitrary nesting of parentheses and negation.

As optimisations we use a deque as an intermediate value before returning a list
so as to avoid copying when inserting on the left.