btrfs: fix subvolume snapshots
Refactor and add option to fall-back to a move not just a copy_path.
Always try to snapshot from workspace to the output directory but
fall-back to using move.

After the removal of OutputFormat.subvolume in favour of
--use-subvolumes taking a BTRFS snapshot for the final output no longer
occurred.

Since snapshots only ever apply to directories lets always call
btrfs_maybe_snapshot_subvolume() and extend it to be able to do a move()
as well as a copy_path() as the fall-back behaviour if subvolumes are
not possible or a snapshot fails.

Fixes: #1569