Set Dir::Etc from config file for apt
Specifying it as an argument to apt-get doesn't work.
from apt.conf(5):

Dir::Etc contains the location of configuration files, sourcelist gives
the location of the sourcelist and main is the default configuration
file (setting has no effect, unless it is done from the config file
specified by the "APT_CONFIG" environment variable).