Ensure we return the same exit code in debug mode
When running in debug mode, we shouldn't return a different exit
code compared to when we run outside of debug mode.

A trivial example is when running the boot or qemu verbs where we
exit with the exit code of the container/VM by raising an instance
of CalledProcessError. In --debug mode, this exception is handled
as an unhandled exception by the python runtime which always exits
with 1 when an unhandled exception is encountered, whereas outside
of debug mode we exit with the value contained in
CalledProcessError.returncode.

With this commit, we'll always exit with the return value in
CalledProcessError.returncode.