Rename btrfs.py to tree.py and move copy_path() into it
Let's hide btrfs subvolumes as an implementation detail in a new
tree module. We have two functions, make_tree() and copy_tree()
(renamed from copy_path()), which will both take advantage of btrfs
subvolumes if configured to do so.

We also extend copy_tree() to not try to snapshot files.